+++ R/atc_property.R
@@ -25,20 +25,20 @@
 #' @param x a (vector of a) valid \code{\link{atc}} code or any text that can be coerced to a valid atc with \code{\link{as.atc}}
 #' @param property one of the column names of one of the \code{\link{antibiotics}} data set, like \code{"atc"} and \code{"official"}
 #' @param language language of the returned text, defaults to English (\code{"en"}) and can be set with \code{\link{getOption}("AMR_locale")}. Either one of \code{"en"} (English) or \code{"nl"} (Dutch).
-#' @rdname ab_property
-#' @return A vector of values. In case of \code{ab_tradenames}, if \code{x} is of length one, a vector will be returned. Otherwise a \code{\link{list}}, with \code{x} as names.
+#' @rdname atc_property
+#' @return A vector of values. In case of \code{atc_tradenames}, if \code{x} is of length one, a vector will be returned. Otherwise a \code{\link{list}}, with \code{x} as names.
 #' @export
 #' @importFrom dplyr %>% left_join pull
 #' @seealso \code{\link{antibiotics}}
 #' @inheritSection AMR Read more on our website!
 #' @examples
-#' ab_atc("amcl")         # J01CR02
-#' ab_name("amcl")        # Amoxicillin and beta-lactamase inhibitor
-#' ab_name("amcl", "nl")  # Amoxicilline met enzymremmer
-#' ab_trivial_nl("amcl")  # Amoxicilline/clavulaanzuur
-#' ab_certe("amcl")       # amcl
-#' ab_umcg("amcl")        # AMCL
-ab_property <- function(x, property = 'official') {
+#' as.atc("amcl")         # J01CR02
+#' atc_name("amcl")        # Amoxicillin and beta-lactamase inhibitor
+#' atc_name("amcl", "nl")  # Amoxicilline met enzymremmer
+#' atc_trivial_nl("amcl")  # Amoxicilline/clavulaanzuur
+#' atc_certe("amcl")       # amcl
+#' atc_umcg("amcl")        # AMCL
+atc_property <- function(x, property = 'official') {
   property <- property[1]
   if (!property %in% colnames(AMR::antibiotics)) {
     stop("invalid property: ", property, " - use a column name of the `antibiotics` data set")
@@ -53,15 +53,9 @@ ab_property <- function(x, property = 'official') {
   )
 }
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_atc <- function(x) {
-  as.character(as.atc(x))
-}
-
-#' @rdname ab_property
-#' @export
-ab_official <- function(x, language = NULL) {
+atc_official <- function(x, language = NULL) {
 
   if (is.null(language)) {
     language <- getOption("AMR_locale", default = "en")[1L]
@@ -69,40 +63,40 @@ ab_official <- function(x, language = NULL) {
     language <- tolower(language[1])
   }
   if (language %in% c("en", "")) {
-    ab_property(x, "official")
+    atc_property(x, "official")
   } else if (language == "nl") {
-    ab_property(x, "official_nl")
+    atc_property(x, "official_nl")
   } else {
     stop("Unsupported language: '", language, "' - use one of: 'en', 'nl'", call. = FALSE)
   }
 }
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_name <- ab_official
+atc_name <- atc_official
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_trivial_nl <- function(x) {
-  ab_property(x, "trivial_nl")
+atc_trivial_nl <- function(x) {
+  atc_property(x, "trivial_nl")
 }
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_certe <- function(x) {
-  ab_property(x, "certe")
+atc_certe <- function(x) {
+  atc_property(x, "certe")
 }
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_umcg <- function(x) {
-  ab_property(x, "umcg")
+atc_umcg <- function(x) {
+  atc_property(x, "umcg")
 }
 
-#' @rdname ab_property
+#' @rdname atc_property
 #' @export
-ab_tradenames <- function(x) {
-  res <- ab_property(x, "trade_name")
+atc_tradenames <- function(x) {
+  res <- atc_property(x, "trade_name")
   res <- strsplit(res, "|", fixed = TRUE)
   if (length(x) == 1) {
     res <- unlist(res)