+++ tests/testthat/test-atc_property.R
@@ -19,20 +19,19 @@
 # Visit our website for more info: https://msberends.gitab.io/AMR.     #
 # ==================================================================== #
 
-context("ab_property.R")
+context("atc_property.R")
 
-test_that("ab_property works", {
-  expect_equal(ab_certe("amox"), "amox")
-  expect_equal(ab_name("amox", language = "en"), "Amoxicillin")
-  expect_equal(ab_name("amox", language = "nl"), "Amoxicilline")
-  expect_equal(ab_official("amox", language = "en"), "Amoxicillin")
-  expect_equal(ab_trivial_nl("amox"), "Amoxicilline")
-  expect_equal(ab_umcg("amox"), "AMOX")
-  expect_equal(class(ab_tradenames("amox")), "character")
-  expect_equal(class(ab_tradenames(c("amox", "amox"))), "list")
-  expect_equal(ab_atc("amox"), as.character(as.atc("amox")))
+test_that("atc_property works", {
+  expect_equal(atc_certe("amox"), "amox")
+  expect_equal(atc_name("amox", language = "en"), "Amoxicillin")
+  expect_equal(atc_name("amox", language = "nl"), "Amoxicilline")
+  expect_equal(atc_official("amox", language = "en"), "Amoxicillin")
+  expect_equal(atc_trivial_nl("amox"), "Amoxicilline")
+  expect_equal(atc_umcg("amox"), "AMOX")
+  expect_equal(class(atc_tradenames("amox")), "character")
+  expect_equal(class(atc_tradenames(c("amox", "amox"))), "list")
 
-  expect_error(ab_property("amox", "invalid property"))
-  expect_error(ab_name("amox", language = "INVALID"))
-  expect_output(print(ab_name("amox", language = NULL)))
+  expect_error(atc_property("amox", "invalid property"))
+  expect_error(atc_name("amox", language = "INVALID"))
+  expect_output(print(atc_name("amox", language = NULL)))
 })